+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/dialogs/LangProjectWizardFirstPage.java
@@ -12,6 +12,7 @@ package melnorme.lang.ide.ui.dialogs;
 
 
 import static melnorme.lang.ide.ui.utils.DialogPageUtils.severityToMessageType;
+import static melnorme.utilbox.core.CoreUtil.list;
 import static org.eclipse.jface.layout.GridDataFactory.fillDefaults;
 
 import org.eclipse.core.resources.IProject;
@@ -47,6 +48,7 @@ import melnorme.util.swt.components.misc.StatusMessageWidget;
 import melnorme.utilbox.concurrency.OperationCancellation;
 import melnorme.utilbox.core.CommonException;
 import melnorme.utilbox.fields.FieldValueListener.FieldChangeListener;
+import melnorme.utilbox.fields.validation.ValidationSource;
 import melnorme.utilbox.status.IStatusMessage;
 import melnorme.utilbox.status.Severity;
 import melnorme.utilbox.status.StatusException;
@@ -122,7 +124,7 @@ public abstract class LangProjectWizardFirstPage extends WizardPage {
 		createContents_ValidationGroups(parent);
 		
 		FieldChangeListener listener = this::validateDialog;
-		nameGroup.textField.addChangeListener(listener);
+		nameGroup.nameField.addChangeListener(listener);
 		locationGroup.addChangeListener(listener);
 		
 		validateDialog();
@@ -143,22 +145,22 @@ public abstract class LangProjectWizardFirstPage extends WizardPage {
 	
 	public static class NameGroup extends CompositeWidget {
 		
-		protected final TextFieldWidget textField = new TextFieldWidget(WizardMessages.LangNewProject_NameGroup_label);
+		protected final TextFieldWidget nameField = new TextFieldWidget(WizardMessages.LangNewProject_NameGroup_label);
 		
 		public NameGroup() {
 			super(true);
-			addChildWidget(textField);
+			addChildWidget(nameField);
 			this.layoutColumns = 2;
 			
-			validation.addFieldValidationX(true, textField, this::validateProjectName);
+			nameField.addFieldValidator2(true, this::getProjectHandleFor);
+		}
+		
+		public TextFieldWidget nameField() {
+			return nameField;
 		}
 		
 		public String getName() {
-			return textField.getFieldValue();
-		}
-		
-		public TextFieldWidget getNameField() {
-			return textField;
+			return nameField.getFieldValue();
 		}
 		
 		public void validateProjectName() throws StatusException {
@@ -166,20 +168,21 @@ public abstract class LangProjectWizardFirstPage extends WizardPage {
 		}
 		
 		public IProject getProjectHandle2() throws StatusException {
-			return new ProjectValidator().getProjectHandle(getName());
+			return getProjectHandleFor(getName());
+		}
+		
+		public IProject getProjectHandleFor(String projectName) throws StatusException {
+			return new ProjectValidator().getProjectHandle(projectName);
 		}
 		
 		/* -----------------  ----------------- */
 		
 		public void postSetFocus() {
-			SWTUtil.post_setFocus(textField.getFieldControl());
+			SWTUtil.post_setFocus(nameField.getFieldControl());
 		}
 		
 	}
 	
-	public IProject getProjectHandle2() throws CommonException {
-		return nameGroup.getProjectHandle2();
-	}
 	
 	/* ----------------- Location ----------------- */
 	
@@ -192,9 +195,10 @@ public abstract class LangProjectWizardFirstPage extends WizardPage {
 				WizardMessages.LangNewProject_Location_Directory_buttonLabel
 			);
 		
-			nameGroup.getNameField().addChangeListener(this::updateDefaultFieldValue);
+			nameGroup.nameField.addChangeListener(this::updateDefaultFieldValue);
 			
-			addFieldValidationX(true, () -> doValidate());
+			getValidation().addFieldValidation2(true, list(nameGroup.nameField, this), 
+				ValidationSource.fromRunnable(this::doValidate));
 		}
 		
 		protected String getProjectName() {
@@ -275,7 +279,7 @@ public abstract class LangProjectWizardFirstPage extends WizardPage {
 		public void updateWidgetFromInput() {
 			IProject projectHandle;
 			try {
-				projectHandle = getProjectHandle2();
+				projectHandle = nameGroup.getProjectHandle2();
 			} catch(CommonException e) {
 				projectHandle = null;
 			}