+++ documentation/ChangeLog.md
@@ -1,7 +1,8 @@
 ## release ChangeLog
 
 ### (NextVersion)
-* Fixed editor NPE when opening files from a non-workspace VCS revision. 
+ * Fixed stale validation message for location field in New Project Wizard.
+ * Fixed editor NPE when opening files from a non-workspace VCS revision. 
 
 ### 0.16.1
  * Updated minimum CDT to version `9.0` (CDT from Eclipse Neon).